Updated Mechanics and added a sway effect
The flashlight can now have a min intensity that can be set. I made function to clean up the code a bit. A sway effect is added and can also be adjusted and tweak. I added an IEnumerator to add a flicker effect. It works but I don't think its perfect.