New Interactions - Pickable Interactable and Basic Inventory Implemented
The Pickable interaction has been implemented. Now players can interact with a pickable interactable to add it to their inventory, and the UI shows the icon of the item picked up. No way to drop items yet, but the methods for removing the icon from the UI and the items from the inventory are there waiting to be used.