New Interactions - Fillable Slot Interactable Implemented
Added a new Slot interactable that can be filled with a specific item. In this case, the Sphere Pickable Object can be picked up and inserted in the Slot Cylinder Object, any other object will be rejected. Once the slot is filled, a switch in the slot data will be turned to true and can be used by other object in the scene to interpret the data, same as the button interactable!